Revert the tbdoc executable Docker container to use the /app dir
Permissions errors are encountered when running the tbdoc utility
as a non-root user on GitHub Action runners. To enable this
utility to process Markdown files for both local development
and during GitHub Actions workflows, commands must be run under
the root user.  To avoid the misperception that Node.js packages
under /home/node/app are run by a non-root user, the container
will be reverted to storing packages under /app.

Signed-off-by: Frank Hinek <frankhinek@users.noreply.github.com>